+++ src/components/Sidebar/Options/Text/SceneText.tsx
@@ -12,7 +12,7 @@ const SceneText: React.FC = () => {
 
     const { text, setText, sceneText, setSceneText, modelWrapper } = scene;
 
-    if (!text || !sceneText) return <p>{t("please-wait")}</p>;
+    if (!text || !sceneText) return <p>{t("loadings.please-wait")}</p>;
 
     const handleSceneTextVisible = (
         event: React.ChangeEvent<HTMLInputElement>,