+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 2026.03.05
 - March 5th trauma (My eyes feels heavy... )
 - Added custom choices text box
+- Updated Model and Background list
+- Updated the tutorial with the new features
 
 # 2026.02.14
 - Added button that re-applies Live2D Parameters after an emotion is applied.